+++ app/src/main/kotlin/my/github/MrxSiN/modeevolved/signal/LocationUpdates.kt
@@ -50,6 +50,6 @@ class LocationUpdates(
     }
 
     private companion object {
-        const val ATTRIBUTION_TAG = "ModeEvolved"
+        const val ATTRIBUTION_TAG = "PixelModesEvolved"
     }
 }